Пройдерны тесты

+++ internal/models/req.go
@@ -1,8 +1,8 @@
 package models
 
 type WithdrawReq struct {
-	OrderId string `json:"order"`
-	Sum     int64  `json:"sum"`
+	OrderId string  `json:"order"`
+	Sum     float32 `json:"sum"`
 }
 
 type UserReq struct {
